+++ src/azure-cli/azure/cli/command_modules/vm/_template_builder.py
@@ -248,7 +248,8 @@ def build_vm_resource(  # pylint: disable=too-many-locals
         os_publisher=None, os_offer=None, os_sku=None, os_version=None, os_vhd_uri=None,
         attach_os_disk=None, os_disk_size_gb=None, custom_data=None, secrets=None, license_type=None, zone=None,
         disk_info=None, boot_diagnostics_storage_uri=None, ultra_ssd_enabled=None, proximity_placement_group=None,
-        computer_name=None, dedicated_host=None, priority=None, max_billing=None, eviction_policy=None):
+        computer_name=None, dedicated_host=None, priority=None, max_billing=None, eviction_policy=None,
+        provision_vm_agent=None):
 
     os_caching = disk_info['os'].get('caching')
 
@@ -277,6 +278,14 @@ def build_vm_resource(  # pylint: disable=too-many-locals
                     ]
                 }
             }
+
+        if provision_vm_agent is not None:
+            if 'linuxConfiguration' not in os_profile:
+                os_profile['linuxConfiguration'] = {}
+            os_profile['linuxConfiguration']['provisionVMAgent'] = provision_vm_agent
+            # if 'windowsConfiguration' not in os_profile:
+            #     os_profile['windowsConfiguration'] = {}
+            # os_profile['windowsConfiguration']['provisionVMAgent'] = provision_vm_agent
 
         if secrets:
             os_profile['secrets'] = secrets